+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+[VERSION 1.0.1]
+- Support to simulate keep frame buffer, saving last render state to a Texture and drawing this texture every frame using Graphics.Blit
+ps: This feature is automatically enabled in iOS when using Metal because this target API will force clear Buffer on every frame.
+
 [VERSION 1.0]
 Initial Version